[#4310] Re-implement spell slots config using ApplicationV2
Re-implements `ActorSpellSlotsConfig` as `SpellSlotsConfig` using
ApplicationV2 while keeping the design the same.
