+++ templates/actors/config/spell-slots-config.hbs
@@ -1,5 +1,4 @@
-<form autocomplete="off">
-    <header>{{ localize "DND5E.SpellSlotsConfig" }}</header>
+<section>
     <ol class="unlist form-grid form-grid-20">
         {{#each overrides}}
         <li>
@@ -10,4 +9,4 @@
         </li>
         {{/each}}
     </ol>
-</form>
+</section>
